[apps] Fix int size in vsetivli

+++ software/apps/riscv-tests/isa/rv32uiv/vsetivli.c
@@ -7,12 +7,12 @@
 #include "vector_macros.h"
 
 void TEST_CASE1(void) {
-  uint64_t avl, vtype, vl;
-  uint64_t vlmul = 0;
-  uint64_t vsew = 0;
-  uint64_t vta = 1;
-  uint64_t vma = 1;
-  uint64_t golden_vtype;
+  uint32_t avl, vtype, vl;
+  uint32_t vlmul = 0;
+  uint32_t vsew = 0;
+  uint32_t vta = 1;
+  uint32_t vma = 1;
+  uint32_t golden_vtype;
   vtype(golden_vtype, vlmul, vsew, vta, vma);
   asm volatile("vsetivli %[AVL], 16, e8, m1, ta, ma" : [AVL] "=r"(avl));
   read_vtype(vtype);
@@ -21,12 +21,12 @@ void TEST_CASE1(void) {
 }
 
 void TEST_CASE2(void) {
-  uint64_t avl, vtype, vl;
-  uint64_t vlmul = 2;
-  uint64_t vsew = 2;
-  uint64_t vta = 0;
-  uint64_t vma = 1;
-  uint64_t golden_vtype;
+  uint32_t avl, vtype, vl;
+  uint32_t vlmul = 2;
+  uint32_t vsew = 2;
+  uint32_t vta = 0;
+  uint32_t vma = 1;
+  uint32_t golden_vtype;
   vtype(golden_vtype, vlmul, vsew, vta, vma);
   asm volatile("vsetivli %[AVL], 31, e32, m4, tu, ma" : [AVL] "=r"(avl));
   read_vtype(vtype);
@@ -36,12 +36,12 @@ void TEST_CASE2(void) {
 
 // Zero avl
 void TEST_CASE3(void) {
-  uint64_t avl, vtype, vl;
-  uint64_t vlmul = 3;
-  uint64_t vsew = 3;
-  uint64_t vta = 0;
-  uint64_t vma = 0;
-  uint64_t golden_vtype;
+  uint32_t avl, vtype, vl;
+  uint32_t vlmul = 3;
+  uint32_t vsew = 3;
+  uint32_t vta = 0;
+  uint32_t vma = 0;
+  uint32_t golden_vtype;
   vtype(golden_vtype, vlmul, vsew, vta, vma);
   asm volatile("vsetivli %[AVL], 0, e16, m8, tu, mu" : [AVL] "=r"(avl));
   read_vtype(vtype);
